remove Swagger() decorator and move Swagger optiions to options of decorators for HTTP methods
