+++ src/controllers/index.ts
@@ -57,5 +57,4 @@ export * from './PATCH';
 export * from './POST';
 export * from './PUT';
 export * from './Serializer';
-export * from './Swagger';
 export * from './TRACE';
